tests: ODTs are represented by GDataDocumentsDocument in Drive v2
GDataDocumentText is bound to application/vnd.google-apps.document,
which represents native Drive-specific text documents. Anything that's
not a Drive-specific type, except PDFs, are meant to be represented as
GDataDocumentsDocument.

Trying to upload an ODT as a GDataDocumentText confuses things. We
pass the ODT's MIME type to the GDataUploadStream, which is not the
MIME type of the associated entry. The updated GDataEntry that is
obtained as a side-effect of the upload is set to match the stream's
MIME type. Therefore it is a GDataDocumentDocuments, which isn't the
same type as the initial entry we started with.

https://bugzilla.gnome.org/show_bug.cgi?id=684920